perf(pflip): C4 — native GPU float atomics in p2g
Replaces the atomicCompSwap-loop emulation of float atomic add with
native atomicAdd via the GL_EXT_shader_atomic_float Vulkan extension.
The CAS loop was correct but every particle deposit triggered a
read-modify-CAS-retry cycle on four u-faces plus four v-faces plus
eight weights per particle; under dense contention (8 particles per
cell) the retry rate was the single largest cost in p2g.

The extension is supported on NVIDIA Maxwell+, AMD GCN3+, Intel Tiger
Lake+, and Apple Silicon via MoltenVK — all the platforms this
prototype targets. If a user runs on hardware without the extension,
shader compile will fail at startup; we'll add a CAS-path fallback if
that ever actually happens.

Changes:
- pflip_p2g.glsl: #extension GL_EXT_shader_atomic_float : require,
  u_vel/v_vel/u_weights/v_weights declared as float data[],
  atomicAddFloatU/V/UW/VW helpers deleted (28 lines gone), direct
  atomicAdd(buffer.data[slot], val) at each scatter site.
- pflip_normalize.glsl: matching float declarations, delete the
  uintBitsToFloat/floatBitsToUint cruft that only existed because p2g
  needed uint for CAS. Shader is now plain arithmetic.
- pflip_clear_grid.glsl: matching float declarations, zero via 0.0.
  Same bit pattern as 0u, but the declaration is now consistent with
  the other two shaders.

Other shaders (pflip_g2p, pflip_save_vel, pflip_extrapolate,
pflip_viscosity, fluid_*) already declared these buffers as float[]
so they need no changes. The underlying storage layout is unchanged
— only p2g's declaration-level interpretation flipped.

Verified: Godot 4.6.2 on Apple M5 Max compiles all three shaders
without error; pflip_test scene runs 300 frames with a 1000-particle
center blob and produces no runtime warnings.

+++ src/shaders/pflip_clear_grid.glsl
@@ -15,19 +15,19 @@ layout(set = 0, binding = 0, std430) restrict buffer Params {
 } params;
 
 layout(set = 0, binding = 1, std430) restrict buffer UVel {
-    uint data[];  // packed float bits, used with atomic compswap
+    float data[];
 } u_vel;
 
 layout(set = 0, binding = 2, std430) restrict buffer VVel {
-    uint data[];
+    float data[];
 } v_vel;
 
 layout(set = 0, binding = 3, std430) restrict buffer UWeights {
-    uint data[];
+    float data[];
 } u_weights;
 
 layout(set = 0, binding = 4, std430) restrict buffer VWeights {
-    uint data[];
+    float data[];
 } v_weights;
 
 layout(set = 0, binding = 5, std430) restrict buffer Density {
@@ -48,12 +48,12 @@ void main() {
     // For simplicity, we dispatch over max(w+1, w) x max(h, h+1) and check bounds per buffer.
 
     if (x <= w && y < h) {
-        u_vel.data[y * (w + 1) + x] = 0u;
-        u_weights.data[y * (w + 1) + x] = 0u;
+        u_vel.data[y * (w + 1) + x] = 0.0;
+        u_weights.data[y * (w + 1) + x] = 0.0;
     }
     if (x < w && y <= h) {
-        v_vel.data[y * w + x] = 0u;
-        v_weights.data[y * w + x] = 0u;
+        v_vel.data[y * w + x] = 0.0;
+        v_weights.data[y * w + x] = 0.0;
     }
     if (x < w && y < h) {
         density.data[y * w + x] = 0u;
